Changed posn store position with doubles(actually did this last edit but forgot to document). Changed move to take in double value.

+++ src/model/ObjectInterfaces/Movable.java
@@ -24,5 +24,5 @@ public interface Movable {
    * @param dx the change in x.
    * @param dy the change in y.
    */
-  void move(int dx, int dy);
+  void move(double dx, double dy);
 }
